Fixed the noise generation for Swift.
For Swiftwe don't have to apply the poisson noise, instead, due to how
the detector is built, to simulate the noise we take a random gaussian
value centered on the value of the "self._model" simulated LC. The sigma
is taken from real data in the following way. We divide the range of LC
in 7 bins of "peak LC", and in each bin we append the values of all the
"errs" of the real LC. Then for each simulated LC, we check the value of
the peak, we choose the corresponding interval, and we sample the sigma
from the errs of the real LCs in that bin of peaks.