UPDATED TO PYGAD3 + solved "stuck" bug
I've updated the code to pygad3 (just add 'ga_instance' argument to the
function 'fitness_function()').
Then I've also solved the bug that caused the code to remain stuck
indefinitely; it was due to the fact that the lower limit for 'tau_max'
was too low, and so it used to generate LC with small T90, which did not
get accepted by the 'apply_constraints()' function.